+++ src/main/java/com/wagglex2/waggle/common/security/handler/LoginSuccessHandler.java
@@ -113,7 +113,6 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
      * @param name Refresh / Access Token Cookie Name
      * @param value Refresh / Access Token 값
      * @param maxAge Refresh / Access 만료 기간
-     * @return
      */
     private ResponseCookie createResponseCookie(String name, String value, Duration maxAge) {
